DEMRUM-4814: extract shared waitUntil test helper
